ref: move route recipes type to service recipes type
