+++ src/services/home/homeRecipesService.ts
@@ -1,9 +1,10 @@
 import { HomeRecipe } from "../../controllers/home/homeRecipesController.js";
 import { generateRecipeKey } from "../recipes/helper.js";
-import { RecipesSimple } from "../../router/v1/recipes/recipes.js";
+
 import { getImgUrl } from "../../utils/img.js";
 import { splitString } from "../../utils/normalize.js";
 import { getRecentRecipes } from "../recipes/recipesSearchService.js";
+import { RecipesSimple } from "../recipes/type.js";
 
 export const homeRecipesService = async () => {
   const result = await getRecentRecipes(18);
