+++ src/services/recipes/recipesCreateService/index.ts
@@ -4,7 +4,7 @@ import mysqlDB from "../../../db/mysql.js";
 
 import { SerializedUser } from "../../../config/passport.js";
 
-import { INewRecipe } from "../../../router/v1/recipes/recipes.js";
+import { INewRecipe } from "../type.js";
 import { generateRecipeKey } from "../helper.js";
 import { CreateRecipeBody } from "../../../controllers/recipes/recipesCreateController.js";
 import {
